% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/action_button.R
\name{action_buttonInput}
\alias{action_buttonInput}
\title{<Add Title>}
\usage{
action_buttonInput(
  inputId,
  label,
  variant = c("contained", "outlined", "text"),
  size = c("medium", "small", "large"),
  disabled = FALSE,
  fontColor = "#ffffff",
  mainColor = "#e05151"
)
}
\description{
<Add Description>
}
