% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkbox_field.R
\name{checkbox_fieldInput}
\alias{checkbox_fieldInput}
\title{<Add Title>}
\usage{
checkbox_fieldInput(
  inputId,
  label,
  default = TRUE,
  labelPlacement = c("end", "start", "top", "bottom"),
  fontColor = "#ffffff",
  mainColor = "#e05151"
)
}
\description{
<Add Description>
}
