% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_field.R
\name{input_fieldInput}
\alias{input_fieldInput}
\title{<Add Title>}
\usage{
input_fieldInput(
  inputId,
  label,
  variant = c("outlined", "filled", "standard"),
  fontColor = "#ffffff",
  mainColor = "#e05151"
)
}
\description{
<Add Description>
}
