% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/button_file_input.R
\docType{data}
\name{RAW_DATA_PATH}
\alias{RAW_DATA_PATH}
\title{<File Input Button>}
\format{
An object of class \code{NULL} of length 0.
}
\usage{
RAW_DATA_PATH
}
\description{
<Add Description>
}
\keyword{datasets}
