% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_field.R
\name{select_fieldInput}
\alias{select_fieldInput}
\title{<Add Title>}
\usage{
select_fieldInput(
  inputId,
  label,
  value = NULL,
  disabled = FALSE,
  secondaryColor = "#ffffff",
  mainColor = "#e05151"
)
}
\description{
<Add Description>
}
