% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_upload.R
\name{source_uploadInput}
\alias{source_uploadInput}
\title{<Add Title>}
\usage{
source_uploadInput(
  inputId,
  label,
  fontColor = "#ffffff",
  mainColor = "#e05151",
  variant = c("outlined", "contained", "text"),
  size = c("medium", "small", "large"),
  filetype_accept = ".txt, .rock, text/yaml",
  startIcon = FALSE,
  endIcon = FALSE
)
}
\description{
<Add Description>
}
