% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textarea.R
\name{textareaInput}
\alias{textareaInput}
\title{<Add Title>}
\usage{
textareaInput(
  inputId,
  label,
  value = NULL,
  rowsInit,
  secondaryColor = "#ffffff",
  mainColor = "#e05151"
)
}
\description{
<Add Description>
}
